requirements.txt
Flask>=2.0
openai-whisper
torch           
pandas
scikit-learn
ffmpeg-python  